Move utils to client directory
